pyproject.toml
[tool.poetry]
name = "tomo"
version = "0.1.0"
description = "A library for topic modeling (tomo)."
license = "MIT"
authors = ["Mayank Nagda <nagdamayank05@gmail.com>"]
readme = "README.md"
packages = [
    {include = "tomo", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
matplotlib = "^3.7.2"
scienceplots = "^2.1.0"
gensim = "^4.3.2"
scikit-learn = "^1.3.0"
pyyaml = "^6.0.1"
pandas = "^2.1.0"
transformers = "^4.33.1"
datasets = "^1.14.0"
tomotopy = "^0.12.5"
seaborn = "^0.12.2"
sentence-transformers = "^2.2.2"
spacy = "^3.6.1"
bertopic = "^0.15.0"
tokenease = "^0.1.5"
wandb = "^0.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
